Crate moonbeam_rpc_trace

trace_filter RPC handler and its associated service task. The RPC handler rely on CacheTask which provides a future that must be run inside a tokio executor.

The implementation is composed of multiple tasks :

Many calls the RPC handler Trace::filter, communicating with the main task.

A main CacheTask managing the cache and the communication between tasks.

For each traced block an async task responsible to wait for a permit, spawn a blocking task and waiting for the result, then send it to the main CacheTask.

Structs

CacheBatchId

An opaque batch ID.

CacheRequester

Allows to interact with the cache task.

CacheTask

Type wrapper for the cache task, generic over the Client, Block and Backend types.

FilterRequest

Trace

RPC handler. Will communicate with a CacheTask through a CacheRequester.

Traits

TraceServer

Server trait implementation for the Trace RPC API.