moonbeam_rpc_trace

Trait TraceServer

pub trait TraceServer:
    Sized
    + Send
    + Sync
    + 'static {
    // Required method
    fn filter<'life0, 'async_trait>(
        &'life0 self,
        filter: FilterRequest,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionTrace>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the Trace RPC API.

Required Methods

fn filter<'life0, 'async_trait>( &'life0 self, filter: FilterRequest, ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionTrace>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.